Rewrite engine, and spin off example into separate repo
The engine should now be much, much clearer: better organised, and better commented.
Additionally, it is now compatible with more recent RGBDS versions.

The auto-linewrapper now uses a "lookahead" (i.e. speculative execution :D)
instead of copying the text stream to RAM; this should improve performance,
in particular removing the CPU usage spikes inherent to this strategy.

The way the engine is meant to be used is also drastically overhauled:
insta-print now yields control one tile at a time, which allows processing
other events in parallel (as in Shock Lobster's mod of gb-vwf)

There are also far fewer variables all around, so the engine is much simpler
to initialise.

A major problem remains to be corrected: glyphs wider than 5 pixels are truncated,
due to the LUT approach used to rotate pixels. This will be corrected Very Soon™